+++ RELEASENOTES.md
@@ -16,6 +16,7 @@
 - Fix incorrect dehacked state lookup for arachnotron RUN7 and RUN8 states. Fixes KDIKDIZD suicide bomber scream.
 - Fix flats in dynamic render path that were flagged not to render.
 - Fix crash when trying to copy text to the clipboard in linux.
+- Fix cases where automap would mark lines that couldn't be seen.
 
 ## Misc:
 - Use DrawArraysInstanced instead of geometry shader for sprite rendering (allows for MacOS support).